refactor(const): decouple protocol vs domain constants and enums
- Relocate protocol payload constants to ramses_tx.const and domain constants/enums to ramses_rf.const and ramses_rf.enums
- Update module import sites across src/ramses_rf, src/ramses_cli, and tests
- Decouple ramses_tx.address from higher-level DEV_TYPE_MAP domain lookup
- Coordinated via ramses-rf/ramses_rf#986 (ref: ramses-rf/ramses_rf#639, ramses-rf/ramses_rf#930, ramses-rf/ramses_cc#884, ramses-rf/ramses_cc#868)
