+++ src/ramses_rf/binding_fsm.py
@@ -17,15 +17,8 @@ import voluptuous as vol
 from ramses_rf.address import Address
 from ramses_rf.commands.builders import build_dto
 from ramses_rf.commands.core import Command as Intent
-from ramses_rf.enums import Action
-from ramses_tx import (
-    ALL_DEVICE_ID,
-    NON_DEVICE_ID,
-    CommandDTO,
-    DevType,
-    Priority,
-    QosParams,
-)
+from ramses_rf.enums import Action, DevType
+from ramses_tx import ALL_DEVICE_ID, NON_DEVICE_ID, CommandDTO, Priority, QosParams
 from ramses_tx.typing import DeviceIdT
 
 from . import exceptions as exc
