+++ src/ramses_rf/devices/hvac_sensors.py
@@ -13,14 +13,13 @@ from ramses_rf.const import (
     SZ_CO2_LEVEL,
     SZ_INDOOR_HUMIDITY,
     SZ_PRESENCE_DETECTED,
-    SZ_TEMPERATURE,
-    Code,
     DevType,
 )
 from ramses_rf.enums import Action
 from ramses_rf.messages import Message
 from ramses_rf.models import DeviceTraits, HvacState
 from ramses_tx import Packet, Priority
+from ramses_tx.const import SZ_TEMPERATURE, Code
 
 from .dev_base import BatteryState, DeviceHvac, Fakeable
