+++ src/ramses_rf/devices/opentherm_bridge.py
@@ -7,7 +7,6 @@ from datetime import timedelta as td
 from typing import Any, Final
 
 from ramses_rf.const import (
-    FC,
     HEARTBEAT_TIMEOUT_OTB,
     SZ_BOILER_OUTPUT_TEMP,
     SZ_BOILER_RETURN_TEMP,
@@ -38,17 +37,15 @@ from ramses_rf.const import (
     SZ_FLAME_ACTIVE,
     SZ_FLAME_SIGNAL_LOW,
     SZ_MAX_REL_MODULATION,
-    SZ_OEM_CODE,
     SZ_OTC_ACTIVE,
     SZ_OUTSIDE_TEMP,
     SZ_REL_MODULATION_LEVEL,
     SZ_SUMMER_MODE,
-    Code,
     DevType,
 )
 from ramses_rf.models import DeviceTraits, OpenThermState
 from ramses_tx import Priority
-from ramses_tx.const import SZ_NUM_REPEATS, SZ_PRIORITY, MsgId
+from ramses_tx.const import FC, SZ_NUM_REPEATS, SZ_OEM_CODE, SZ_PRIORITY, Code, MsgId
 from ramses_tx.typing import PayDictT
 
 from ..protocol.opentherm import OtDataId
