+++ src/ramses_rf/messages/base.py
@@ -11,11 +11,12 @@ from typing import TYPE_CHECKING, Any, TypeAlias, TypeVar
 
 from ramses_rf.address import Address, id_to_address
 from ramses_tx import CommandDTO, PacketDTO
+from ramses_tx.const import SZ_DHW_IDX, SZ_DOMAIN_ID, SZ_UFH_IDX, SZ_ZONE_IDX
 from ramses_tx.models import DeviceId, RawPacket, TransportMessage
 from ramses_tx.typing import DeviceIdT
 
 from .. import exceptions as exc
-from ..const import DEV_TYPE_MAP, SZ_DHW_IDX, SZ_DOMAIN_ID, SZ_UFH_IDX, SZ_ZONE_IDX
+from ..const import DEV_TYPE_MAP
 from ..parsers.decoder import decode_packet
 from ..protocol.ramses import CODE_IDX_ARE_COMPLEX
 from ..routing import RoutingContext, StateHeader
