+++ src/ramses_rf/parsers/heating.py
@@ -10,11 +10,11 @@ import logging
 from datetime import datetime as dt, timedelta as td
 from typing import TYPE_CHECKING, Any
 
+from ramses_rf.const import DEV_ROLE_MAP, DEV_TYPE_MAP, ZON_MODE_MAP, ZON_ROLE_MAP
+from ramses_rf.enums import DevRole
 from ramses_tx import exceptions as exc
 from ramses_tx.address import hex_id_to_dev_id
 from ramses_tx.const import (
-    DEV_ROLE_MAP,
-    DEV_TYPE_MAP,
     F9,
     FA,
     FC,
@@ -44,9 +44,6 @@ from ramses_tx.const import (
     SZ_ZONE_IDX,
     SZ_ZONE_MASK,
     SZ_ZONE_TYPE,
-    ZON_MODE_MAP,
-    ZON_ROLE_MAP,
-    DevRole,
 )
 from ramses_tx.helpers import (
     hex_to_bool,
