+++ src/ramses_rf/pipeline/ingestion.py
@@ -18,13 +18,9 @@ from ramses_rf.address import HGI_DEV_ADDR, Address
 from ramses_rf.commands.builders import build_dto
 from ramses_rf.commands.core import Command as Intent_
 from ramses_rf.const import (
-    SZ_ACTIVE,
     SZ_ACTUATOR_COUNTDOWN,
-    SZ_ACTUATOR_ENABLED,
     SZ_AIR_QUALITY,
     SZ_AIR_QUALITY_BASIS,
-    SZ_BATTERY_LEVEL,
-    SZ_BATTERY_LOW,
     SZ_BYPASS_MODE,
     SZ_BYPASS_POSITION,
     SZ_BYPASS_STATE,
@@ -34,56 +30,34 @@ from ramses_rf.const import (
     SZ_CO2_LEVEL,
     SZ_COOL_ACTIVE,
     SZ_CYCLE_COUNTDOWN,
-    SZ_DATETIME,
     SZ_DEWPOINT_TEMP,
     SZ_DHW_ACTIVE,
     SZ_DHW_FLOW_RATE,
-    SZ_DIFFERENTIAL,
-    SZ_DOMAIN_ID,
     SZ_EXHAUST_FAN_SPEED,
     SZ_EXHAUST_FLOW,
     SZ_EXHAUST_TEMP,
     SZ_FAN_INFO,
     SZ_FAN_MODE,
     SZ_FAN_RATE,
-    SZ_FLAME_ON,
-    SZ_HEAT_DEMAND,
     SZ_INDOOR_HUMIDITY,
     SZ_INDOOR_TEMP,
-    SZ_LANGUAGE,
     SZ_MAX_REL_MODULATION,
     SZ_MINUTES,
-    SZ_MODE,
-    SZ_MODULATION_LEVEL,
-    SZ_NAME,
     SZ_OUTDOOR_HUMIDITY,
     SZ_OUTDOOR_TEMP,
-    SZ_OVERRUN,
     SZ_POST_HEAT,
     SZ_PRE_HEAT,
     SZ_PRESENCE_DETECTED,
-    SZ_PRESSURE,
     SZ_REL_MODULATION_LEVEL,
-    SZ_RELAY_DEMAND,
-    SZ_RELAY_FAILSAFE,
     SZ_REMAINING_DAYS,
     SZ_REMAINING_MINS,
     SZ_REMAINING_PERCENT,
     SZ_REQ_REASON,
     SZ_REQ_SPEED,
-    SZ_SETPOINT,
-    SZ_SETPOINT_BOUNDS,
     SZ_SPEED_CAPABILITIES,
     SZ_SUPPLY_FAN_SPEED,
     SZ_SUPPLY_FLOW,
     SZ_SUPPLY_TEMP,
-    SZ_SYSTEM_MODE,
-    SZ_TEMPERATURE,
-    SZ_UFH_IDX,
-    SZ_UNTIL,
-    SZ_WINDOW_OPEN,
-    SZ_ZONE_IDX,
-    Code,
 )
 from ramses_rf.enums import Action
 from ramses_rf.messages import Message
@@ -103,6 +77,34 @@ from ramses_rf.models import (
 )
 from ramses_rf.protocol.opentherm import OtDataId
 from ramses_rf.state_projector import _get_dhw_zone_from_msg, _route_2411_to_fan
+from ramses_tx.const import (
+    SZ_ACTIVE,
+    SZ_ACTUATOR_ENABLED,
+    SZ_BATTERY_LEVEL,
+    SZ_BATTERY_LOW,
+    SZ_DATETIME,
+    SZ_DIFFERENTIAL,
+    SZ_DOMAIN_ID,
+    SZ_FLAME_ON,
+    SZ_HEAT_DEMAND,
+    SZ_LANGUAGE,
+    SZ_MODE,
+    SZ_MODULATION_LEVEL,
+    SZ_NAME,
+    SZ_OVERRUN,
+    SZ_PRESSURE,
+    SZ_RELAY_DEMAND,
+    SZ_RELAY_FAILSAFE,
+    SZ_SETPOINT,
+    SZ_SETPOINT_BOUNDS,
+    SZ_SYSTEM_MODE,
+    SZ_TEMPERATURE,
+    SZ_UFH_IDX,
+    SZ_UNTIL,
+    SZ_WINDOW_OPEN,
+    SZ_ZONE_IDX,
+    Code,
+)
 
 # --- Translation Maps (Static Constant Blocks) ---
