+++ src/ramses_rf/state/entity_state.py
@@ -17,11 +17,10 @@ from typing import TYPE_CHECKING, Any, Literal, overload
 from ramses_tx.address import ALL_DEVICE_ID
 
 # noqa: F401, isort: skip, pylint: disable=unused-import
-from ramses_tx.const import I_, RP, RQ, Code, VerbT
+from ramses_tx.const import I_, RP, RQ, SZ_DOMAIN_ID, SZ_NAME, SZ_ZONE_IDX, Code, VerbT
 from ramses_tx.typing import PayDictT
 
 from .. import exceptions as exc
-from ..const import SZ_DOMAIN_ID, SZ_NAME, SZ_ZONE_IDX
 from ..messages import ApplicationMessage, Message
 from ..protocol.ramses import CODES_SCHEMA
 from ..routing import RoutingContext, StateHeader
