+++ src/ramses_rf/state_projector.py
@@ -10,17 +10,30 @@ import uuid
 from collections.abc import Callable
 from typing import TYPE_CHECKING, Any, Final
 
+from ramses_tx.const import (
+    SZ_ACTIVE,
+    SZ_DATETIME,
+    SZ_DIFFERENTIAL,
+    SZ_HEAT_DEMAND,
+    SZ_LANGUAGE,
+    SZ_MODE,
+    SZ_OVERRUN,
+    SZ_RELAY_DEMAND,
+    SZ_SETPOINT,
+    SZ_SYSTEM_MODE,
+    SZ_TEMPERATURE,
+    SZ_UNTIL,
+    Code,
+)
+
 from . import exceptions as exc, quirks
 from .const import (
-    SZ_ACTIVE,
     SZ_AIR_QUALITY,
     SZ_AIR_QUALITY_BASIS,
     SZ_BYPASS_MODE,
     SZ_BYPASS_POSITION,
     SZ_BYPASS_STATE,
     SZ_CO2_LEVEL,
-    SZ_DATETIME,
-    SZ_DIFFERENTIAL,
     SZ_EXHAUST_FAN_SPEED,
     SZ_EXHAUST_FLOW,
     SZ_EXHAUST_TEMP,
@@ -30,33 +43,23 @@ from .const import (
     SZ_FILTER_DIRTY,
     SZ_FROST_CYCLE,
     SZ_HAS_FAULT,
-    SZ_HEAT_DEMAND,
     SZ_INDOOR_HUMIDITY,
     SZ_INDOOR_TEMP,
-    SZ_LANGUAGE,
     SZ_MINUTES,
-    SZ_MODE,
     SZ_OUTDOOR_HUMIDITY,
     SZ_OUTDOOR_TEMP,
-    SZ_OVERRUN,
     SZ_POST_HEAT,
     SZ_PRE_HEAT,
     SZ_PRESENCE_DETECTED,
-    SZ_RELAY_DEMAND,
     SZ_REMAINING_DAYS,
     SZ_REMAINING_MINS,
     SZ_REMAINING_PERCENT,
     SZ_REQ_REASON,
     SZ_REQ_SPEED,
-    SZ_SETPOINT,
     SZ_SPEED_CAPABILITIES,
     SZ_SUPPLY_FAN_SPEED,
     SZ_SUPPLY_FLOW,
     SZ_SUPPLY_TEMP,
-    SZ_SYSTEM_MODE,
-    SZ_TEMPERATURE,
-    SZ_UNTIL,
-    Code,
 )
 from .devices.hvac_ventilators import HvacVentilator
 from .messages import Message
