+++ src/ramses_rf/topology.py
@@ -26,8 +26,10 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
 
+from ramses_tx.const import F9, FA, FC, FF, SZ_ACTUATORS, SZ_SENSOR
+
 from . import exceptions as exc
-from .const import F9, FA, FC, FF, SZ_ACTUATORS, SZ_SENSOR, DevType
+from .enums import DevType
 from .schemas import SZ_CIRCUITS
