+++ tests/tests/test_apis_heat.py
@@ -9,13 +9,12 @@ from typing import Any
 from ramses_rf.address import Address
 from ramses_rf.commands.builders import build_dto
 from ramses_rf.commands.core import Command as Intent
-from ramses_rf.const import SZ_DOMAIN_ID
 from ramses_rf.enums import Action
 from ramses_rf.helpers import shrink
 from ramses_rf.messages import Message
 from ramses_rf.parsers.helpers import parse_fault_log_entry
 from ramses_tx.address import HGI_DEV_ADDR
-from ramses_tx.const import SZ_TIMESTAMP
+from ramses_tx.const import SZ_DOMAIN_ID, SZ_TIMESTAMP
 from ramses_tx.dtos import CommandDTO as Command
 from ramses_tx.packet import Packet
 from ramses_tx.typing import DeviceIdT
